extract AtomEdit and Edit into new ra_text_edit crate
