+++ crates/ra_text_edit/src/text_utils.rs
@@ -1,0 +1,5 @@
+use text_unit::{TextRange, TextUnit};
+
+pub fn contains_offset_nonstrict(range: TextRange, offset: TextUnit) -> bool {
+    range.start() <= offset && offset <= range.end()
+}
